requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.1
mysqlclient==2.2.7
packaging==26.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
